requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
watchdog>=3.0.0
filelock>=3.12.0
slowapi>=0.1.9
pyyaml>=6.0
python-dateutil>=2.8.0
